Switch QuokkADB back to PlatformIO
Moved QuokkADB back to the PlatformIO building process from CMake.
This time using a custom spin of the earlephilhower arduinopico
framework that includes USB host support.

The biggest code change to support this is using a wrapper around
Arduino Serial library for ADBuino and for QuokkADB using
the same wrapper interface but using pico/printf as the underlying
serial logging method.