Only respond to talk reg 3 after flush
The previous code would cause the ADB driver to think
there was a mouse and keyboard on all addresses.
This was due to the mouse and keyboard responding to talk reg 3
before a flush was sent to the device.

Now the devices will only respond to a to talk reg 3 after a reset
and a flush has occurred. Which the ADB driver now only displays
the one two devices.