Some good ideas, not working with Gravis gamepad
There are some positive fixes in the commit, but breaks when a Gravis
gamepad is on the ADB bus.

It sometimes enumerates the devices the same as the Wombat,
but then it queries all the addresses between mouse moves.
In another case it seems to work but the enumeration is off.
The keyboard gets moved to address 0xF and the Gravis gamepad
is at address 0x2 (keyboard's default address).