Fix right mouse button and QuokkADB freezing
The right mouse button wasn't lifting Control after being lifted.
The fix for this also seems to fix the QuokkADB freezing after a while.
The freezing only occurred in the release build. Under the
debug build it worked fine.

A possible memory leak was also addressed in the simple circular queue.
Pointers to objects get the `delete` if the queue is full.